Fix chain id from pay rate merge

+++ libraries/blockchain/include/bts/blockchain/genesis_config.hpp
@@ -9,7 +9,7 @@ namespace bts { namespace blockchain {
   {
      std::string        name;
      public_key_type    owner;
-     uint8_t            delegate_pay_rate = 255;
+     int                delegate_pay_rate = 255;
   };
 
   struct asset_config // these are all market-issued assets
